Update database description to reflect changes to Librus API
Librus increased its announcements' id lengths making them too long for the database column.
This resulted in re-posting the entire page's contents every time the script was run.
It also revealed that the old facebook sdk for php broke when the page was eventually blocked
from posting.

This commit changes the column specification in the databases.ini.example, changes the
facebook sdk to a newer version and fixes a minor bug where the script wouldn't purge the
post from the database when it was removed from the page as a result of abruptly stopping
the script.